% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sheet.R
\name{build_datasheet}
\alias{build_datasheet}
\title{Build an SVG datasheet from somnofy, axivity and EMA data}
\usage{
build_datasheet(somnofy, axivity, ema, output_file, format = "pdf")
}
\arguments{
\item{somnofy}{Data frame containing Somnofy sleep data}

\item{axivity}{Data frame containing Axivity sleep data}

\item{ema}{Data frame containing EMA data}

\item{output_file}{Path to save the generated datasheet (pdf format)}
}
\value{
None. The function saves the datasheet SVG file to the specified location.
}
\description{
Build an SVG datasheet from somnofy, axivity and EMA data
}
