% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{create_burst_labels}
\alias{create_burst_labels}
\title{Create burst labels for plotting}
\usage{
create_burst_labels(df, burst_col = "burst", night_col = "night")
}
\arguments{
\item{df}{A nocturn sessions dataframe}

\item{burst_col}{The name of the column indicating bursts (default is "burst")}
}
\value{
The input data frame with an additional \code{burst_label} column
}
\description{
Burst labels show the date range for each burst in the data (typically EMA or Axivity).
}
\details{
This function needs the \code{night} column to be defined
}
