% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_datasheet.R
\name{generate_datasheet}
\alias{generate_datasheet}
\title{Generate a datasheet for a given participant, using data from the Ambient-BD datastore}
\usage{
generate_datasheet(datastore_root, participant_id, output_file, format = "pdf")
}
\arguments{
\item{datastore_root}{Root directory of the Ambient-BD datastore}

\item{participant_id}{ID of the participant (e.g., "ABD001")}

\item{output_file}{Path to save the generated datasheet SVG file}
}
\value{
None. The function saves the datasheet SVG file to the specified location.
}
\description{
Generate a datasheet for a given participant, using data from the Ambient-BD datastore
}
