% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{preprocess_df}
\alias{preprocess_df}
\title{Rename columns to differentiate between methods (somnofy, ema, axivity)}
\usage{
preprocess_df(df, colnames, method_name)
}
\description{
Rename columns to differentiate between methods (somnofy, ema, axivity)
}
