% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{shortest_time_diff}
\alias{shortest_time_diff}
\title{Returns the shortest difference between two times (in minutes)}
\usage{
shortest_time_diff(t1, t2)
}
\description{
Returns the shortest difference between two times (in minutes)
}
